optimised code
Optimized All pages for responsiveness using Bootstrap, improved code organization, and added meta description and keywords for SEO."

This commit description highlights the main changes you made to the code, including:

Using Bootstrap to make the page more responsive and improve the layout on different devices and screen sizes.
Organizing the code in a more structured and readable way, making it easier to maintain and modify in the future.
Adding meta description and keywords tags to the page's head section to improve its search engine optimization (SEO) and help search engines understand the content of the page.
Favicon replaced and removed default React logo and icons.